add support for using sysparm_display_value=all
Currently the use of sysparm_display_value=all breaks the paging logic which assumes that the property value found via the 'orderByField' will be something we can append a "Z" to then DateTimeOffset.Parse.

This fails when you're using sysparm_display_value=all since the value of the orderByField will be a JObject with properties of value and display_value so you end up with an exception during the query similar to this:

String '{
  "display_value": "2024-03-04 21:15:20",
  "value": "2024-03-04 21:15:20"
}Z' was not recognized as a valid DateTime.

This changes the parsing to use DateTimeStyles.AssumeUniversal instead of appending a "Z" and to fall back to DateTimeOffset.MinValue for failure cases instead of throwing an exception.  Since it throws an exception, it makes it difficult for query callers that require using sysparm_display_value=all